+++ src/main/java/miniswing/pilot/web/dto/OrdersUpdateRequestDto.java
@@ -8,12 +8,12 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class OrdersUpdateRequestDto {
     //private Integer svc_mgmt_num;
-    private Integer acnt_num;
+    private long acnt_num;
     private String prod_id;
     private String svc_scrb_dt;
 
     @Builder
-    public OrdersUpdateRequestDto(Integer acnt_num, String prod_id, String svc_scrb_dt) {
+    public OrdersUpdateRequestDto(long acnt_num, String prod_id, String svc_scrb_dt) {
         //this.svc_mgmt_num = svc_mgmt_num;
         this.acnt_num = acnt_num;
         this.prod_id = prod_id;
